refactored KeyCache
