+++ test/source/platform/store.ts
@@ -24,20 +24,4 @@ export class Store {
     return [];
   }
 
-  public static decryptedKeyCacheSet = (k: OpenPGP.key.Key) => {
-    // tests don't need this
-  }
-
-  public static decryptedKeyCacheGet = (longid: string): OpenPGP.key.Key | undefined => {
-    return undefined; // tests don't need this
-  }
-
-  public static armoredKeyCacheSet = (armored: string, k: OpenPGP.key.Key) => {
-    // tests don't need this
-  }
-
-  public static armoredKeyCacheGet = (armored: string): OpenPGP.key.Key | undefined => {
-    return undefined; // tests don't need this
-  }
-
 }
